Fix some test
