+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/admin/impl/TransactionsBase.java
@@ -88,7 +88,8 @@ public abstract class TransactionsBase extends AdminResource {
                 });
             }
         } else {
-            asyncResponse.resume(new RestException(503, "Broker don't support transaction!"));
+            asyncResponse.resume(new RestException(503,
+                    "This Broker is not configured with transactionCoordinatorEnabled=true."));
         }
     }
 }
